Renamed the top-level npg namespace to npg_porch.
npg.porch is collapsed into npg_porch
npg.porchdb is reorganised into npg_porch.db
The entry point of the application is relocated and
renamed npg/porch_server.py -> npg_porch/server.py

Import statements in teh code and tests and docs are
updated.
