+++ src/npg_porch/server.py
@@ -21,7 +21,7 @@
 from fastapi import FastAPI
 from fastapi.responses import HTMLResponse
 
-from npg.porch.endpoints import pipelines, tasks
+from npg_porch.endpoints import pipelines, tasks
 
 #https://fastapi.tiangolo.com/tutorial/bigger-applications/
 #https://fastapi.tiangolo.com/tutorial/metadata
